Separate monthlyreturns and to-date returns. Aggregate queries into their own file and export. Start migrating financial calcs to util folder
