+++ src/components/Routes/Fund/FundOverview/FundFactSheet/FundMonthlyReturnTable.tsx
@@ -9,6 +9,7 @@ import { calculateReturn } from '~/utils/finance';
 import { Block } from '~/storybook/Block/Block';
 import { Spinner } from '~/storybook/Spinner/Spinner.styles';
 import { SectionTitle } from '~/storybook/Title/Title';
+import { useFetchMonthlyFundPrices } from './FundMetricsQueries';
 
 export interface MonthlyReturnTableProps {
   address: string;
@@ -35,32 +36,15 @@ interface DisplayData {
   price: BigNumber;
 }
 
-async function fetchMonthlyFundPrices(key: string, address: string) {
-  const url = process.env.MELON_METRICS_API;
-  const queryAddress = `${url}/api/monthend?address=${address}`;
-  const response = await fetch(queryAddress)
-    .then((response) => response.json())
-    .catch((error) => console.log(error));
-  return response;
-}
-
-function useFetchMonthlyFundPrices(fund: string) {
-  const address = React.useMemo(() => fund.toLowerCase(), [fund]);
-  return useQuery(['prices', address], fetchMonthlyFundPrices, {
-    refetchOnWindowFocus: false,
-  });
-}
-
-export default function FundMonthlyReturnTable(props: MonthlyReturnTableProps) {
+export const FundMonthlyReturnTable: React.FC<MonthlyReturnTableProps> = ({ address }) => {
   const today = React.useMemo(() => new Date(), []);
   const fund = useFund();
   const fundInception = fund.creationTime!;
   const activeMonths = differenceInCalendarMonths(today, fundInception);
   const inactiveMonths = differenceInCalendarMonths(fundInception, startOfYear(today));
-  console.log('inactive months: ', inactiveMonths);
 
   const { data: fundMonthlyData, error: fundMonthlyError, isFetching: fundMonthlyFetching } = useFetchMonthlyFundPrices(
-    props.address
+    address
   );
 
   if (!fundMonthlyData || fundMonthlyFetching) {
@@ -71,14 +55,11 @@ export default function FundMonthlyReturnTable(props: MonthlyReturnTableProps) {
     );
   }
 
-  // create two arrays of objects and append one to the other. first one is all months in 2020 prior to fund inception.
-  // find length of that array by declaring the beginning of the year, then taking difference in calendar months
-  // put the month on an object with the corresponding return
   const activeMonthReturns: DisplayData[] = fundMonthlyData.data.map(
     (item: DepthTimelineItem, index: number, arr: DepthTimelineItem[]) => {
       if (index === 0) {
         return {
-          price: calculateReturn(new BigNumber(1), new BigNumber(item.calculations.price)),
+          price: calculateReturn(new BigNumber(item.calculations.price), new BigNumber(1)),
           date: format(item.timestamp * 1000, 'MMM yyy'),
         };
       }
@@ -92,20 +73,18 @@ export default function FundMonthlyReturnTable(props: MonthlyReturnTableProps) {
     }
   );
 
-  const inactiveMonthReturns: DisplayData[] = new Array(inactiveMonths + 1)
+  const inactiveMonthReturns: DisplayData[] = new Array(inactiveMonths)
     .fill(null)
     .map((item, index: number) => {
-      return { date: format(subMonths(today, index + activeMonths), 'MMM yyy'), price: new BigNumber('n/a') };
+      return { date: format(subMonths(today, index + activeMonths + 1), 'MMM yyy'), price: new BigNumber('n/a') };
     })
     .reverse();
 
   const displayDataset = inactiveMonthReturns.concat(activeMonthReturns);
-  // it'll almost always be the case that there are more months than there are returns
 
   return (
     <Block>
       <SectionTitle>Monthly Returns (Share Price)</SectionTitle>
-
       <Table>
         <tbody>
           <HeaderRow>
@@ -116,7 +95,7 @@ export default function FundMonthlyReturnTable(props: MonthlyReturnTableProps) {
           <BodyRow>
             {displayDataset.map((item, index) => (
               <BodyCell key={index}>
-                <FormattedNumber suffix={'%'} value={item.price} decimals={2} />
+                <FormattedNumber suffix={'%'} value={item.price} decimals={2} colorize={true} />
               </BodyCell>
             ))}
           </BodyRow>
@@ -124,4 +103,98 @@ export default function FundMonthlyReturnTable(props: MonthlyReturnTableProps) {
       </Table>
     </Block>
   );
+};
+
+/**
+ *  Index query logic that we'll need later:
+ *
+ *   const [historicalMonthlyIndexPrices, setHistoricalMonthlyIndexPrices] = React.useState<BigNUmber[][] | undefined>(undefined)
+ *
+ *  const fundMonthDates = new Array(ageInMonths + 1)
+ *    .fill(null)
+ *    .map(item, index: number) => {
+ *      const targetMonth = subMonths(today, index);
+ *      return [startOfMonth(targetMonth), endOfMonth(targetMonth)] as [Date, Date];
+ *    })
+ *      .reverse()
+ *
+ *  React.useMemo(async () => {
+ *    const prices = await fetchMultipleIndexPrices(fundMonthDates)
+ *    setHistoricalMonthlyIndexPrices(prices)
+ *  }, [fund])
+ *
+ *  const { data: indexData, error: indexError, isFetching: indexFetching } = useFetchIndexPrices(
+ *    indexQueryStartDate,
+ *    endOfDay(today).toISOString()
+ *  ) // where indexQueryStartDate is something like => startOfDay(subDays(subWeeks(today, 1), 1)).toISOString();
+ *
+ */
+
+/**
+ * RETURN MATH FOR VOLATILITY AND VAR CALS
+ */
+// // indexData is an array of bignumbers, every index price since the date that's passed
+// // indexDailyReturns is an array of bignumbers, with index 0 being the return on day 0 and index(len-1) the return yesterday
+// const indexDailyReturns = indexData ? calculateDailyLogReturns(indexData) : [new BigNumber('NaN')];
+// // periodIndexVol takes the daily returns as a parameter and returns a bignumber == the volatility of those returns
+// const periodIndexVol = indexData
+//   ? calculateVolatility(calculateStdDev(indexDailyReturns), indexDailyReturns.length)
+//   : new BigNumber('Nan');
+// // periodIndexReturn shows the return if one were to hold an index for the time period
+// const periodIndexReturn = indexData ? calculatePeriodReturns(indexData) : new BigNumber('Nan');
+// // periodIndexVar takes the daily returns as a parameter and returns an object where lowZ == 95% CI and highZ = 99% CI
+// const periodIndexVAR = indexDailyReturns && calculateVAR(indexDailyReturns);
+
+// // fundData.data is an array of price objects with onchain and offchain prices as well as fund holdings data
+// // onChainPriceUpdates is an array of BigNumbers with the duplicate onchain prices stripped out (would otherwise remain constant for a day)
+// const onChainPriceUpdates = fundData && stripDuplicateOnchainPrices(fundData.data);
+// // onChainDailyReturns is an array of BigNumbers showing the inter-day returns starting with day 1's return  and ending with yesterday's
+// const onChainDailyReturns = onChainPriceUpdates
+//   ? calculateDailyLogReturns(onChainPriceUpdates)
+//   : [new BigNumber('Nan')];
+// // periodFundVol takes the onChainDailyReturns as a parameter and returns a bignumber === the volatility of those returns
+// const periodFundVol =
+//   onChainDailyReturns && calculateVolatility(calculateStdDev(onChainDailyReturns), onChainDailyReturns.length);
+// // periodFundReturn takes the onChainPriceUpdates as a parameter and returns a bignumner
+// const periodFundReturn = onChainPriceUpdates ? calculatePeriodReturns(onChainPriceUpdates) : new BigNumber('Nan');
+// // periodFundVar takes the daily returns as a parameter and returns an object where lowZ = 95% CI and highZ = 9% CI
+// const periodFundVAR = onChainDailyReturns && calculateVAR(onChainDailyReturns);
+
+{
+  /* <DictionaryEntry>
+        <DictionaryLabel>{depth} Fund Volatility </DictionaryLabel>
+        <DictionaryData textAlign={'right'}>
+          <FormattedNumber decimals={2} value={periodFundVol} suffix={'%'} />
+        </DictionaryData>
+      </DictionaryEntry>
+      <DictionaryEntry>
+        <DictionaryLabel>{depth} BitWise 10 Index Vol</DictionaryLabel>
+        <DictionaryData textAlign={'right'}>
+          <FormattedNumber decimals={2} value={periodIndexVol} suffix={'%'} />
+        </DictionaryData>
+      </DictionaryEntry>
+      <DictionaryEntry>
+        <DictionaryLabel>{depth} Fund Return</DictionaryLabel>
+        <DictionaryData textAlign={'right'}>
+          <FormattedNumber decimals={2} value={periodFundReturn} colorize={true} suffix={'%'} />
+        </DictionaryData>
+      </DictionaryEntry>
+      <DictionaryEntry>
+        <DictionaryLabel>{depth} BitWise 10 Index Return</DictionaryLabel>
+        <DictionaryData textAlign={'right'}>
+          <FormattedNumber decimals={2} value={periodIndexReturn} colorize={true} suffix={'%'} />
+        </DictionaryData>
+      </DictionaryEntry>
+      <DictionaryEntry>
+        <DictionaryLabel>{depth} Fund VAR</DictionaryLabel>
+        <DictionaryData textAlign={'right'}>
+          <FormattedNumber decimals={2} value={periodFundVAR?.lowZ} colorize={false} suffix={'%'} />
+        </DictionaryData>
+      </DictionaryEntry>
+      <DictionaryEntry>
+        <DictionaryLabel>{depth} BitWise 10 Index VAR</DictionaryLabel>
+        <DictionaryData textAlign={'right'}>
+          <FormattedNumber decimals={2} value={periodIndexVAR?.lowZ} colorize={false} suffix={'%'} />
+        </DictionaryData>
+      </DictionaryEntry> */
 }
